Move layer controls into separate component and render as modal
