+++ css/layer-controls.less
@@ -1,0 +1,20 @@
+.map-layer-controls{
+  .overlay-button{
+    font-size:12px;
+  }
+}
+
+.map-layer-content{
+    display: block;
+    width: 200px;
+    height: 150px;
+    position: absolute;
+    left: 150px;
+    bottom: 35px;
+    z-index: 3;
+    padding-top:20px;
+    > div + div{
+      display:block;
+      margin:10px;
+    }
+}
