+++ js/containers/bottom-controls.js
@@ -3,6 +3,7 @@ import pureRender from 'pure-render-decorator'
 import {connect} from 'react-redux'
 import * as actions from '../actions'
 import AnimationButtons from '../components/animation-buttons'
+import LayerControls from './layer-controls'
 import Slider from 'rc-slider'
 import ccLogoSrc from '../../images/cc-logo.png'
 import screenfull from 'screenfull'
@@ -67,13 +68,9 @@ class BottomControls extends Component {
     }
     this.handleTimeRange = this.handleTimeRange.bind(this)
     this.handleMagRange = this.handleMagRange.bind(this)
-    this.handleBaseLayerChange = this.handleBaseLayerChange.bind(this)
-    this.handlePlateLayerChange = this.handlePlateLayerChange.bind(this)
     this.handleAnimStep = this.handleAnimStep.bind(this)
     this.handlePlayPauseBtnClick = this.handlePlayPauseBtnClick.bind(this)
     this.handleResetBtnClick = this.handleResetBtnClick.bind(this)
-    this.handleEarthquakeLayerChange = this.handleEarthquakeLayerChange.bind(this)
-    this.handleVolcanoLayerChange = this.handleVolcanoLayerChange.bind(this)
   }
 
   componentDidMount() {
@@ -96,26 +93,6 @@ class BottomControls extends Component {
     setFilter('maxMag', value[1])
   }
 
-  handleBaseLayerChange(event) {
-    const {setBaseLayer} = this.props
-    const layer = event.target.value
-    setBaseLayer(layer)
-    log('MapLayerChanged', {layer})
-  }
-
-  handlePlateLayerChange(event) {
-    const {setPlatesVisible} = this.props
-    const visible = event.target.checked
-    setPlatesVisible(visible)
-    log('PlatesVisibilityChanged', {visible})
-  }
-
-  handleVolcanoLayerChange(event) {
-    const {setVolcanosVisible} = this.props
-    const visible = event.target.checked
-    setVolcanosVisible(visible)
-    log('VolcanosVisibilityChanged', {visible})
-  }
 
   handleAnimStep(newValue) {
     const {filters, setFilter, setAnimationEnabled} = this.props
@@ -138,12 +115,6 @@ class BottomControls extends Component {
     log('ResetClicked')
   }
 
-  handleEarthquakeLayerChange(event) {
-    const {setEarthquakesVisible} = this.props
-    const visible = event.target.checked
-    setEarthquakesVisible(visible)
-    log("show earthquakes", {visible})
-  }
 
   get dateMarks() {
     const {filters} = this.props
@@ -158,10 +129,6 @@ class BottomControls extends Component {
       Object.assign(marks, sliderTickFormatter(min, max))
     }
     return marks
-  }
-
-  get mapLayerOptions() {
-    return layerInfo.map((m, idx) => <option key={idx} value={m.type}>{m.name}</option>)
   }
 
   get animSpeed() {
@@ -203,31 +170,13 @@ class BottomControls extends Component {
           <div>
             <img src={ccLogoSrc}/>
           </div>
-          <div title="Change the map rendering style">
-            Map type
-            <select value={layers.get('base') } onChange={this.handleBaseLayerChange}>
-              {this.mapLayerOptions}
-            </select>
-          </div>
-          {mode !== '3d' &&
-          <div title="Show Plate Boundaries Overlay">
-            <input type='checkbox' checked={layers.get('plates') } onChange={this.handlePlateLayerChange}
-                   id='plate-border-box'/>
-            <label htmlFor='plate-border-box'>Plate boundaries</label>
-          </div>
-          }
-          <div title="Show Volcanos">
-            <input type='checkbox' checked={layers.get('volcanos') } onChange={this.handleVolcanoLayerChange}
-                   id='volcano-box'/>
-            <label htmlFor='volcano-box'>Volcanos</label>
-          </div>
+         <LayerControls/>
           <div>
             <div className='mag-label'>Magnitudes from <strong>{minMag.toFixed(1)}</strong> to <strong>{maxMag.toFixed(1)}</strong></div>
             <div className='mag-slider'>
               <Slider range min={0} max={10} step={0.1} value={[minMag, maxMag]} onChange={this.handleMagRange}
                       onAfterChange={logMagSliderChange} marks={{0: 0, 5: 5, 10: 10}}/>
             </div>
-            <div className='toggle-earthquakes' title="Show or hide all earthquakes on the map"><input type="checkbox" id="earthquake-toggle" checked={layers.get('earthquakes')} onChange={this.handleEarthquakeLayerChange} /><label htmlFor='earthquake-toggle'>Show Earthquakes</label></div>
             <div className='stats'>
               <span>Currently displaying <strong>{earthquakesCountVisible}</strong> of <strong>{earthquakesCount}</strong> earthquakes </span>
               {magFilter && <span>starting from magnitude <strong>{magnitudeCutOff}</strong>. Zoom in to see weaker earthquakes.</span>}
